pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "global-webscraper"
version = "0.1.0"
description = "Universal Web Scraper & Gemini Enrichment Platform"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.41.0",
    "playwright-stealth>=1.0.6",
    "beautifulsoup4>=4.12.2",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "google-generativeai>=0.3.2",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "W", "C90", "I", "N", "UP", "B", "A", "C4", "DTZ", "T20", "Q"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]